# routes/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GBI API Documentation",
        "description": "Details on all apis for GBI Website will be listed below.",
        "contact": {
            "email": "admin@gbinternational.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "GBI API Server"
        }
    ],
    "paths": {
        "/posts/all/{size}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get list of blogs",
                "description": "Returns list of blogs, paginated by page size mentioned",
                "operationId": "7fe76ce8f295e9e15eae9cd9cf4f86d9",
                "parameters": [
                    {
                        "name": "size",
                        "in": "path",
                        "description": "Pagination size per page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/posts": {
            "post": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Store new blog",
                "description": "Stores new data to the database and returns the stored data",
                "operationId": "8f1e3f3af3d8349b00e557f3a4080c7c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/posts/{id}": {
            "get": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Get blog information",
                "description": "Returns blog data of the blog id provided.",
                "operationId": "046d80bab84334c846bd87f2b5c977bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Blog Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Update existing Blog",
                "description": "Updates blog data of the blog id provided and returns the updated data",
                "operationId": "adf27e587d287e9c3d3d08bc1f339712",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Blog Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Blogs"
                ],
                "summary": "Delete existing Blog",
                "description": "Deletes a record and returns no content",
                "operationId": "f119ad72a9b11a932274b5510a251ab0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Blog Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Itineraries",
            "description": "API Endpoints of Itineraries"
        },
        {
            "name": "Blogs",
            "description": "API Endpoints of Blogs"
        }
    ],
    "security": [
        []
    ]
}